Le coordinate di un centroide del cluster devono essere double non interi

+++ src/main/java/it/sabd2425/gc25client/data/Centroid.java
@@ -4,25 +4,25 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class Centroid {
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
     private final int count;
 
     @JsonCreator
     public Centroid(
-        @JsonProperty("x") int x,
-        @JsonProperty("y") int y,
+        @JsonProperty("x") double x,
+        @JsonProperty("y") double y,
         @JsonProperty("count") int count) {
         this.x = x;
         this.y = y;
         this.count = count;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
